requirements.txt
pytesseract
pdf2image
python-docx
Pillow
opencv-python-headless
